Move ecdsa functions from encoder.ts to own file.
